Move EOL and isWindows to fs/constants.ts, rename EOL to EOL_
