+++ path/mod.ts
@@ -4,7 +4,7 @@
 import * as _win32 from "./win32.ts";
 import * as _posix from "./posix.ts";
 
-import { isWindows } from "./constants.ts";
+import { isWindows } from "../fs/mod.ts";
 
 const path = isWindows ? _win32 : _posix;
 
@@ -24,7 +24,7 @@ export const parse = path.parse;
 export const sep = path.sep;
 export const delimiter = path.delimiter;
 
-export { EOL, SEP, SEP_PATTERN, isWindows } from "./constants.ts";
+export { SEP, SEP_PATTERN } from "./constants.ts";
 export * from "./interface.ts";
 export * from "./glob.ts";
 export * from "./globrex.ts";
