+++ testing/runner.ts
@@ -1,8 +1,8 @@
 #!/usr/bin/env -S deno -A
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
 import { parse } from "../flags/mod.ts";
-import { ExpandGlobOptions, expandGlob } from "../fs/mod.ts";
-import { isWindows, join } from "../path/mod.ts";
+import { ExpandGlobOptions, expandGlob, isWindows } from "../fs/mod.ts";
+import { join } from "../path/mod.ts";
 import { RunTestsOptions, runTests } from "./mod.ts";
 const { DenoError, ErrorKind, args, cwd, exit } = Deno;
